+++ lib/togls/feature_toggle_registry.rb
@@ -7,7 +7,7 @@ module Togls
   # respective entities. This plays a significant portion in the primary DSL as
   # well.
   class FeatureToggleRegistry
-    def initialize
+    def initialize(&block)
       @toggle_repository_drivers = [
         Togls::ToggleRepositoryDrivers::InMemoryDriver.new,
         Togls::ToggleRepositoryDrivers::EnvOverrideDriver.new]
@@ -22,12 +22,7 @@ module Togls
         @toggle_repository_drivers, @feature_repository, @rule_repository)
       @rule_repository.store(Togls::Rules::Boolean.new(true))
       @rule_repository.store(Togls::Rules::Boolean.new(false))
-    end
-
-    def self.create(&block)
-      feature_toggle_registry = new
-      feature_toggle_registry.instance_eval(&block)
-      feature_toggle_registry
+      self.instance_eval(&block) if block_given?
     end
 
     def expand(&block)