+++ spec/togls/feature_toggle_registry_spec.rb
@@ -135,35 +135,23 @@ describe Togls::FeatureToggleRegistry do
       expect(rule_repository).to receive(:store).with(boolean_true_rule)
       subject
     end
-  end
 
-  describe ".create" do
-    it "creates a new instance of a feature toggle registry" do
-      registry = double('registry')
-      expect(Togls::FeatureToggleRegistry).to receive(:new).and_return(registry)
-      b = Proc.new {}
-      Togls::FeatureToggleRegistry.create(&b)
-    end
-
-    context "when block given" do
-      it "calls instance eval with the passed block" do
-        registry = double('registry')
-        allow(Togls::FeatureToggleRegistry).to receive(:new).and_return(registry)
+    context 'when given a block' do
+      it "creates a new instance of a feature toggle registry" do
         b = Proc.new {}
-        expect(registry).to receive(:instance_eval).and_yield(&b)
-        Togls::FeatureToggleRegistry.create(&b)
+        Togls::FeatureToggleRegistry.new(&b)
       end
-    end
 
-    it "returns a configured feature registry object" do
-      b = Proc.new {}
-      expect(Togls::FeatureToggleRegistry.create(&b)).to be_a(Togls::FeatureToggleRegistry)
+      it "returns a configured feature registry object" do
+        b = Proc.new {}
+        expect(Togls::FeatureToggleRegistry.new(&b)).to be_a(Togls::FeatureToggleRegistry)
+      end
     end
   end
 
   describe "#expand" do
     it "instance evals the provided block" do
-      registry = Togls::FeatureToggleRegistry.create do
+      registry = Togls::FeatureToggleRegistry.new do
         feature(:foo, "some description").on
       end
 
@@ -174,7 +162,7 @@ describe Togls::FeatureToggleRegistry do
     end
 
     it "returns the feature toggle repository" do
-      registry = Togls::FeatureToggleRegistry.create do
+      registry = Togls::FeatureToggleRegistry.new do
         feature(:foo, "some description").on
       end